Add component filtered img list
